move device sync under  general tab
